Add ClrStackFrame.Thread
Ensure that you can get to the containing ClrThread from ClrStackFrame.
