+++ src/Microsoft.Diagnostics.Runtime/Desktop/v45runtime.cs
@@ -626,7 +626,7 @@ namespace Microsoft.Diagnostics.Runtime.Desktop
             return data.token;
         }
 
-        protected override DesktopStackFrame GetStackFrame(int res, ulong ip, ulong framePtr, ulong frameVtbl)
+        protected override DesktopStackFrame GetStackFrame(DesktopThread thread, int res, ulong ip, ulong framePtr, ulong frameVtbl)
         {
             DesktopStackFrame frame;
             StringBuilder sb = new StringBuilder();
@@ -647,18 +647,18 @@ namespace Microsoft.Diagnostics.Runtime.Desktop
                         innerMethod = DesktopMethod.Create(this, mdData);
                 }
 
-                frame = new DesktopStackFrame(this, framePtr, frameName, innerMethod);
+                frame = new DesktopStackFrame(this, thread, framePtr, frameName, innerMethod);
             }
             else
             {
                 ulong md;
                 if (_sos.GetMethodDescPtrFromIP(ip, out md) >= 0)
                 {
-                    frame = new DesktopStackFrame(this, ip, framePtr, md);
+                    frame = new DesktopStackFrame(this, thread, ip, framePtr, md);
                 }
                 else
                 {
-                    frame = new DesktopStackFrame(this, ip, framePtr, 0);
+                    frame = new DesktopStackFrame(this, thread, ip, framePtr, 0);
                 }
             }
 
@@ -716,6 +716,7 @@ namespace Microsoft.Diagnostics.Runtime.Desktop
             // Skip size and header
             dataPtr += (ulong)(IntPtr.Size * 2);
 
+            DesktopThread thread = null;
             for (int i = 0; i < (int)count; ++i)
             {
                 ulong ip, sp, md;
@@ -726,7 +727,10 @@ namespace Microsoft.Diagnostics.Runtime.Desktop
                 if (!ReadPointer(dataPtr + (ulong)(2 * IntPtr.Size), out md))
                     break;
 
-                result.Add(new DesktopStackFrame(this, ip, sp, md));
+                if (i == 0)
+                    thread = (DesktopThread)GetThreadByStackAddress(sp);
+
+                result.Add(new DesktopStackFrame(this, thread, ip, sp, md));
 
                 dataPtr += (ulong)elementSize;
             }
