+++ src/Microsoft.Diagnostics.Runtime/Microsoft.Diagnostics.Runtime.csproj
@@ -119,6 +119,7 @@
     <Compile Include="ISymbolResolver.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ClrException.cs" />
+    <Compile Include="Native\NativeThread.cs" />
     <Compile Include="Native\NativeDac.cs" />
     <Compile Include="Native\NativeHeap.cs" />
     <Compile Include="Native\NativeModule.cs" />
